requirements.txt
fpdf2
pytest
tabulate
requests
dotenv